Use public macro to decide whether to use rb_ext_resolve_symbol()
Follow-up for #12. Version detection missed some range of development
versions that are 3.4.0dev, but ship a `Digest` that doesn't
export rb_digest_wrap_metadata(). For example,
https://github.com/Shopify/ruby-definitions/blob/main/rubies/3.4.0-pshopify-preview2

Ideally, we would call rb_digest_make_metadata() that Digest ships
directly instead of reimplementing it, but we already impose an
incredible amount of build burden on from-source consumers.